processing-pipeline/pyproject.toml
[project]
name = "warehouse-processing-pipeline"
version = "0.1.0"
description = "Real-time log processing pipeline for warehouse operations"
authors = [{name = "Warehouse Team", email = "team@warehouse.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.0.2",
    "redis>=4.5.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "prometheus-client>=0.17.0",
    "psutil>=5.9.0",
    "schedule>=1.2.0",
    "aioredis>=2.0.0",
    "asyncio-mqtt>=0.13.0",
    "influxdb-client>=1.38.0",
    "elasticsearch>=8.9.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true